make it compile and run on mac again

+++ include/daScript/simulate/aot.h
@@ -616,7 +616,7 @@ namespace das {
         }
     };
 
-    template <typename TT, uint32_t size>
+    template <typename TT, int size>
     struct TDim {
         using THIS_TYPE = TDim<TT, size>;
         enum { capacity = size };
@@ -631,43 +631,43 @@ namespace das {
     // index
         __forceinline TT & operator () ( int32_t index, Context * __context__ ) {
             uint32_t idx = uint32_t(index);
-            if ( idx>=size ) __context__->throw_error_ex("index out of range, %u of %u", idx, size);
+            if ( idx>=uint32_t(size) ) __context__->throw_error_ex("index out of range, %u of %u", idx, size);
             return data[index];
         }
         __forceinline const TT & operator () ( int32_t index, Context * __context__ ) const {
             uint32_t idx = uint32_t(index);
-            if ( idx>=size ) __context__->throw_error_ex("index out of range, %u of %u", idx, size);
+            if ( idx>=uint32_t(size) ) __context__->throw_error_ex("index out of range, %u of %u", idx, size);
             return data[index];
         }
         __forceinline TT & operator () ( uint32_t idx, Context * __context__ ) {
-            if ( idx>=size ) __context__->throw_error_ex("index out of range, %u of %u", idx, size);
+            if ( idx>=uint32_t(size) ) __context__->throw_error_ex("index out of range, %u of %u", idx, size);
             return data[idx];
         }
         __forceinline const TT & operator () ( uint32_t idx, Context * __context__ ) const {
-            if ( idx>=size ) __context__->throw_error_ex("index out of range, %u of %u", idx, size);
+            if ( idx>=uint32_t(size) ) __context__->throw_error_ex("index out of range, %u of %u", idx, size);
             return data[idx];
         }
     // safe index
         static __forceinline TT * safe_index ( THIS_TYPE * that, int32_t index, Context * ) {
             if (!that) return nullptr;
             uint32_t idx = uint32_t(index);
-            if (idx >= size) return nullptr;
+            if (idx >= uint32_t(size)) return nullptr;
             return that->data + index;
         }
         static __forceinline const TT * safe_index ( const THIS_TYPE * that, int32_t index, Context * ) {
             if (!that) return nullptr;
             uint32_t idx = uint32_t(index);
-            if ( idx>=size ) return nullptr;
+            if ( idx>=uint32_t(size) ) return nullptr;
             return that->data + index;
         }
         static __forceinline TT * safe_index ( THIS_TYPE * that, uint32_t idx, Context * ) {
             if (!that) return nullptr;
-            if ( idx>=size ) return nullptr;
+            if ( idx>=uint32_t(size) ) return nullptr;
             return that->data + idx;
         }
         static __forceinline const TT * safe_index ( const THIS_TYPE * that, uint32_t idx, Context * ) {
             if (!that) return nullptr;
-            if ( idx>=size ) return nullptr;
+            if ( idx>=uint32_t(size) ) return nullptr;
             return that->data + idx;
         }
     };
@@ -676,7 +676,7 @@ namespace das {
         return bitset[ch>>5] & (1u<<uint32_t(ch));
     }
 
-    template <typename TT, uint32_t size>
+    template <typename TT, int size>
     struct das_cast< TDim<TT,size> > {
         template <typename QQ>
         static __forceinline TDim<TT,size> & cast ( const QQ * expr ) {
@@ -684,7 +684,7 @@ namespace das {
         }
     };
 
-    template <typename TT, uint32_t size>
+    template <typename TT, int size>
     struct das_cast< const TDim<TT,size> > {
         template <typename QQ>
         static __forceinline const TDim<TT,size> & cast ( const QQ * expr ) {
